Revert "Add initial effort to baseline and comparison pages"
Part of epic #12142 Baseline v1

This reverts commit a52922dfd1b031e32c5959f9385701940322a822.
Initial effort computation is more tricky than expected.
Some edge cases have been identified lately:
- initial effort can be given by a computed field
- do not sum initial effort of parent artefact with child artifact

So, we decided not to show this field rather than showing strange
values.

+++ include/Domain/BaselineArtifact.php
@@ -36,7 +36,7 @@ class BaselineArtifact
     /** @var string|null */
     private $description;
 
-    /** @var float|null */
+    /** @var int|null */
     private $initial_effort;
 
     /** @var string|null */
@@ -61,7 +61,7 @@ class BaselineArtifact
         int $id,
         ?string $title,
         ?string $description,
-        ?float $initial_effort,
+        ?int $initial_effort,
         ?string $status,
         Project $project,
         int $tracker_id,
@@ -94,7 +94,7 @@ class BaselineArtifact
         return $this->description;
     }
 
-    public function getInitialEffort(): ?float
+    public function getInitialEffort(): ?int
     {
         return $this->initial_effort;
     }
